Such CMake. Much Comments. Wow.

+++ Core/include/Types.h
@@ -38,8 +38,8 @@ typedef jdouble d64; //double type
 
 // PC integer types
 #if defined(OPIFEX_OS64)
-typedef long    i64;
-typedef unsigned long  ui64;  //long type
+typedef long long   i64;
+typedef unsigned long long  ui64;  //long type
 typedef int     i32;
 typedef unsigned int   ui32; //int type
 #elif defined(OPIFEX_OS32)
